removed uneeded changes, added 1 node to pipeline

+++ incasem/gunpowder/merge_labels.py
@@ -28,7 +28,7 @@ class MergeLabels(gp.BatchFilter):
             self,
             classes: Dict[gp.ArrayKey, int],
             output_array: gp.ArrayKey,
-            dtype: Optional[str] = 'uint8',
+            dtype: Optional[str] = 'uint32',
             ambiguous_labels: Optional[str] = 'background'):
 
         self.classes = classes
